refactor: lift BlockingContextGuard + run_async into tahoma-runner
The async-bridge helper was duplicated in dist_spec.rs. Move it into
tahoma-runner so sparse-moe (and any future engine running under
spawn_blocking) shares the same dispatch logic: pick block_on when
we're on a dedicated blocking thread, block_in_place + block_on when
we're on a tokio worker.

On Windows this saves ~60 ms / frame at K2.6 numbers because the
relay loop wraps three async calls per step and block_in_place is
~20 ms vs ~80 ns for the bare block_on path.

dist_spec.rs keeps a private file-scope alias so its many in-file
call sites don't all need updating.
