+++ crates/tahoma-engine-sparse-moe/Cargo.toml
@@ -15,6 +15,7 @@ tahoma-types = { workspace = true }
 tahoma-engine = { workspace = true }
 tahoma-ov-genai-shim = { workspace = true }
 tahoma-int4-gemm = { workspace = true }
+tahoma-runner = { workspace = true }
 tahoma-transport = { workspace = true }
 async-trait = { workspace = true }
 futures = { workspace = true }
